Reorganise dataset layers
