--- ocf_data_sampler/numpy_sample/common_types.py
@@ -1,11 +1,0 @@
-"""This module defines type aliases for numpy and torch data structures used in the project."""
-
-
-from typing import TypeAlias
-
-import numpy as np
-import torch
-
-NumpySample: TypeAlias = dict[str, np.ndarray]
-NumpyBatch: TypeAlias = dict[str, np.ndarray]
-TensorBatch: TypeAlias = dict[str, torch.Tensor]
