+++ ocf_data_sampler/select/spatial_slice.py
@@ -4,7 +4,7 @@ import numpy as np
 import xarray as xr
 
 from ocf_data_sampler.select.geospatial import find_coord_system
-from ocf_data_sampler.select.location import Location
+from ocf_data_samplefrom ocf_data_sampler.spatial import Location
 
 
 def _get_pixel_index_location(da: xr.DataArray, location: Location) -> tuple[int, int]:
