--- ocf_data_sampler/torch_datasets/utils/spatial_slice_for_dataset.py
@@ -1,102 +1,0 @@
-"""Functions for selecting data around a given location."""
-
-
-from ocf_data_sampler.config import Configuration
-from ocf_data_sampler.select import get_indices_in_sorted_unique
-from ocf_data_sampler.select.location import Location
-from ocf_data_sampler.select.spatial_slice import (
-    select_spatial_slice_pixels,
-    select_spatial_slice_pixels_multiple,
-)
-
-
-def slice_datasets_by_space(
-    datasets_dict: dict,
-    location: Location,
-    config: Configuration,
-) -> dict:
-    """Slice the dictionary of input data sources around a given location.
-
-    Args:
-        datasets_dict: Dictionary of the input data sources
-        location: The location to sample around
-        config: Configuration object.
-    """
-    if not set(datasets_dict.keys()).issubset({"nwp", "sat", "generation"}):
-        raise ValueError(
-            "'datasets_dict' should only contain keys 'nwp', 'sat', 'generation'",
-        )
-
-    sliced_datasets_dict = {}
-
-    if "nwp" in datasets_dict:
-        sliced_datasets_dict["nwp"] = {}
-
-        for nwp_key, nwp_config in config.input_data.nwp.items():
-            sliced_datasets_dict["nwp"][nwp_key] = select_spatial_slice_pixels(
-                datasets_dict["nwp"][nwp_key],
-                location,
-                height_pixels=nwp_config.image_size_pixels_height,
-                width_pixels=nwp_config.image_size_pixels_width,
-            )
-
-    if "sat" in datasets_dict:
-        sat_config = config.input_data.satellite
-
-        sliced_datasets_dict["sat"] = select_spatial_slice_pixels(
-            datasets_dict["sat"],
-            location,
-            height_pixels=sat_config.image_size_pixels_height,
-            width_pixels=sat_config.image_size_pixels_width,
-        )
-
-    if "generation" in datasets_dict:
-
-        location_ids = datasets_dict["generation"].location_id.values
-        loc_index = get_indices_in_sorted_unique(location_ids, location.id)
-
-        sliced_datasets_dict["generation"] = datasets_dict["generation"].isel(location_id=loc_index)
-
-    return sliced_datasets_dict
-
-
-def reduce_spatial_extent_of_datasets(
-    datasets_dict: dict,
-    locations: list[Location],
-    config: Configuration,
-) -> dict:
-    """Reduce the spatial extent of the datasets to only cover the locations.
-
-    Args:
-        datasets_dict: Dictionary of the input data sources
-        locations: List of locations to reduce to
-        config: Configuration object
-    """
-    sliced_datasets_dict = {}
-
-    if "nwp" in datasets_dict:
-        sliced_datasets_dict["nwp"] = {}
-
-        for nwp_key, nwp_config in config.input_data.nwp.items():
-            sliced_datasets_dict["nwp"][nwp_key] = select_spatial_slice_pixels_multiple(
-                datasets_dict["nwp"][nwp_key],
-                locations,
-                height_pixels=nwp_config.image_size_pixels_height,
-                width_pixels=nwp_config.image_size_pixels_width,
-            )
-
-
-    if "sat" in datasets_dict:
-        sat_config = config.input_data.satellite
-
-        sliced_datasets_dict["sat"] = select_spatial_slice_pixels_multiple(
-            datasets_dict["sat"],
-            locations,
-            height_pixels=sat_config.image_size_pixels_height,
-            width_pixels=sat_config.image_size_pixels_width,
-        )
-
-    if "generation" in datasets_dict:
-        sliced_datasets_dict["generation"] = datasets_dict["generation"]
-
-    return sliced_datasets_dict
